+++ implementation/js/src/infrastructure/lib+-/marionette/template-cache.js
@@ -4,7 +4,7 @@
  * @author Tim Lauv
  * @created 2014.02.25
  * @updated 2016.03.24
- * @updated 2017.03.02
+ * @updated 2017.05.24
  */
 ;(function(app){
 
@@ -12,14 +12,22 @@
 		// Get the specified template by id.
 		// retrieves the cached tpl obj and load the compiled/text version
 		get: function(templateId, asHTMLText) {
+			if(!templateId)
+				return this.templateCaches;
 		    var cachedTemplate = this.templateCaches[templateId] || this.make(templateId);
 		    return cachedTemplate.load(asHTMLText); //-> cache.loadTemplate()
 		},
 
 		//+ split out a make cache function from the original mono get()
-		//used in a. app.inject.tpl/app.Util.Tpl.remote
-		//consulted in b. cache.loadTemplate
+		//used in app.Util.Tpl (template-builder.js)
 		make: function(templateId, rawTemplate) {
+
+			if(!templateId)
+				throw new Error('DEV::TemplateCache.make() empty templateId: ' + templateId);
+
+			if(this.templateCaches[templateId])
+				console.warn('DEV::TemplateCache.make() overriden::Template ', templateId, ' --> ', rawTemplate);
+			
 			var cachedTemplate = new Marionette.TemplateCache(templateId);
 			this.templateCaches[templateId] = cachedTemplate;
 			cachedTemplate.rawTemplate = rawTemplate;
@@ -57,7 +65,7 @@
 		//1 Override the default raw-template retrieving method 
 		//(invoked by M.TemplateCache.get() by cache.load() if the cache doesn't have cache.compiledTemplate)
 		//We allow both #id or @*.html/.md(remote) and template html string(or string array) as parameter.
-		//This method is only invoked with a template cache miss. So clear your cache if you have changed the template. (app.Util.Tpl.cache.clear(name))
+		//This method is only invoked with a template cache miss. So clear your cache if you have changed the template. (app.Util.Tpl.Cache.clear(name))
 		loadTemplate: function(idOrTplString){
 			//local in-DOM template
 			if(_.string.startsWith(idOrTplString, '#'))